# doc/api.yaml
openapi: 3.0.0
info:
  title: WaSAPhoto API
  version: 1.0.0
  description: |
    **WaSAPhoto - Share Special Moments with Friends**

    **Project Overview:** \
    WaSAPhoto is a social media application designed to keep you connected with your friends and loved ones by enabling you to share special moments through photos. With WaSAPhoto, you can seamlessly upload photos directly from your PC and make them visible to everyone who follows you. 

    **Functional Design Specifications:**

    1. **Photo Stream:**
       - Each user is presented with a dynamic stream of photos in reverse chronological order.
       - The stream provides essential information, including the date and time of each photo's upload, and the number of likes and comments it has received.
       - The photo stream is curated with images from users you are following.

    2. **Interactions:**
       - Users can like photos uploaded by other users and remove their likes at any time.
       - Commenting on photos is encouraged, and users can add comments to any image, even if they uploaded it themselves. Only the authors have the authority to remove their comments.

    3. **User Control:**
       - Users have the ability to ban other users. When a user is banned, they won't have access to any information about the user who banned them. However, the user who initiated the ban can choose to remove it at any time.

    4. **User Profiles:**
...
    
    5. **User Actions:**
...

    6. **User Search:**
...

...

...

...

...
servers:
...
paths:
...
  /users/{uid}/profile:
    parameters:
      - name: uid
        in: path
        required: true
        description: the unique ID hooked to a user.
        schema: { $ref: '#/components/schemas/userID' }

    get:
      security:
        - bearerAuth: []
      tags:
        - "user"
      operationId: getUserProfile
      summary: get user profile
      description: |
        allows getting user's profile information passing the uid.
        The return values will be all the user information and his upload post stream in reverse chronological order
      responses:
        '200':
          description: |
            User information correctly recovered from the server
          content:
            application/json:
              schema:
                description: |
                  server returns the entire profile of a user, 
                  including all his personal information and
                  all the post uploaded.
                type: object
                properties:
                  profile_info:
                    $ref: '#/components/schemas/profileinfo'
                  uploaded_posts:
                    description: contains all the post as array of post object.
                    type: array
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/post'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: '#/components/responses/UnauthorizedError' }
        "404":
          description: |
            the searched uid seems not exists
          content:
            application/json:
              schema:
                description: server returns a not found message
                type: object
                properties:
                  error:
                    description: verbose message error description returned by server
                    type: string
                    minLength: 0
                    maxLength: 256
                    pattern: '^.*?$'
                    example: uid not found
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{uid}/mystream:
    parameters:
      - name: uid
        in: path
        required: true
        description: the unique ID hooked to a user.
        schema: { $ref: '#/components/schemas/userID' }

    get:
      security:
        - bearerAuth: []
      tags:
        - "user"
        - "post"
      operationId: getMyStream
      summary: get user stream photos.
      description: |
        Allows getting user stream photos passing the uid.
        The stream consists in an array of post.
        For getting a binary image it's necessary using the 'Get Image API'
      responses:
        '200':
          description: |
            User stream photo correctly recovered from the server
          content:
            application/json:
              schema:
                description: |
                  server returns an array of post object
                type: object
                properties:
                  posts:
                    description: each object is a post objects.
                    type: array
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/post'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: '#/components/responses/UnauthorizedError' }
        "404":
          description: |
            the searched uid seems not exists
          content:
            application/json:
              schema:
                description: server returns a not found message
                type: object
                properties:
                  error:
                    description: verbose message error description returned by server
                    type: string
                    minLength: 0
                    maxLength: 256
                    pattern: '^.*?$'
                    example: uid not found
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
    username:
      title: username
      description: is the nickname choosed by a user
      type: string
      pattern: '^[A-Za-z0-9]{3,20}$'
      example: Simone01
      minLength: 3
      maxLength: 20
    userID:
      title: the User ID
      description: the userID as an integer. Each user has it own uid.
      type: integer
      minimum: 1
      example: 100
    commentid:
      title: the comment ID
      description: |
        each comment submitted has it own id.
      type: integer
      minimum: 1
      example: 100
    user:
      title: user
      description: all the information about user
      type: object
      properties:
        user_id:
          $ref: '#/components/schemas/userID'
        username:
          $ref: '#/components/schemas/username'
    postid:
      title: post ID
      description: the post ID as an integer
      type: integer
      minimum: 1
      example: 13244
...
    comment:
      title: comment under a post
      description: represents a comment under a photo pubblished
      type: object
      properties:
        id:
          $ref: '#/components/schemas/commentid'
        uid:
          $ref: '#/components/schemas/userID'
        postid:
          $ref: '#/components/schemas/postid'
        message:
          title: message written
          description: the effective message of the comment
          type: string
          minLength: 1
          maxLength: 256
          pattern: ^[a-zA-Z0-9.,!?;:'"\s]+$
          example: a cr4zy comment!
        comment_datetime:
          title: datetime posted comment
          description: |
            represents the date and the time of a post comment according
            to format YYYY-MM-DD HH:MM:SS
          type: string
          minLength: 19
          maxLength: 19
          format: date-time
          example: 2017-07-21 17:32:28
    post:
      title: post content
      description: |
        a single post content, that includes the id of the image,
        the numbers of like and all the comments.
      type: object
      properties:
        postid:
          $ref: '#/components/schemas/postid'
        uid:
          $ref: '#/components/schemas/userID'
        likes:
          title: number of likes
          description: represents the like number of the photo
          type: integer
          minimum: 0
          example: 20
        comments:
          title: list of comments
          description: all the comments under a post in reverse chronological order
          type: array
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/comment'
        upload_datetime:
          title: datetime post uploading
          description: |
            represents the date and the time of a post uploading according
            to format YYYY-MM-DD HH:MM:SS
          type: string
          minLength: 19
          maxLength: 19
          format: date-time
          example: 2017-07-21 17:32:28
    profileinfo:
      description: consists in all the personal information of the user.
      type: object
      properties:
        user:
          $ref: '#/components/schemas/user'
        numpost:
          description: is the amount of all the post uploaded by the user.
          title: the number of uploaded post
          type: integer
          minimum: 0
          example: 10
        follower:
          description: is the number of follower of the user.
          title: the number of follower
          type: integer
          minimum: 0
          example: 10
        following:
          description: is the number of account followed by the user.
          title: the number of followed
          type: integer
          minimum: 0
          example: 10
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            description: |
              server returns a descriptive error message
            properties:
              error:
                description: verbose message error description returned by server
                type: string
                minLength: 0
                maxLength: 256
                pattern: '^.*?$'
                example: log to do action
    BadRequest:
      description: |
        The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema:
            type: object
            description: |
              server returns a descriptive error message
            properties:
              error:
                description: verbose message error description returned by server
                type: string
                minLength: 0
                maxLength: 256
                pattern: '^.*?$'
                example: not correct format
    InternalServerError:
      description: |
        The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: |
              server returns a descriptive error message
            properties:
              error:
                description: verbose message error description returned by server
                type: string
                minLength: 0
                maxLength: 256
                pattern: '^.*?$'
                example: something wrong